model_download2.py
import torch
from modelscope import snapshot_download, AutoModel, AutoTokenizer
import os

model_dir = snapshot_download('OpenBMB/MiniCPM-2B-sft-fp32', cache_dir='/home/zjlab/models', revision='master')
